desktop/requirements.txt
requests>=2.32.0
openpyxl>=3.1.0
Pillow>=10.4.0
pyinstaller>=6.10.0
